All the classes have been updated.
Firstly, I have been clarified it and removed the attribute additionalClasses, because we can simulate it by wrapping the fields into a class which wraps the render of each field in a div.

Secondly, I have created two new classes and refactored the old class RadioField which now it's named RadioButton.

Finally, I have added the attribute value into the render of each field in order to achieve the FormView displaying the data when we need to update a certain model.